requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Model Architecture
timm>=0.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Metrics and Visualization
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Tracking
tqdm>=4.65.0

# Testing
pytest>=7.4.0
hypothesis>=6.82.0
